refactored switch to if/else statement in TestDatabaseActivity.java file

+++ src/com/example/classlocus/test/TestDatabaseActivity.java
@@ -38,21 +38,20 @@ public class TestDatabaseActivity extends ListActivity {
     @SuppressWarnings("unchecked")
     ArrayAdapter<Comment> adapter = (ArrayAdapter<Comment>) getListAdapter();
     Comment comment = null;
-    switch (view.getId()) {
-    case R.id.add:
+    int viewId = view.getId();
+    
+    if (viewId == R.id.add) {
       String[] comments = new String[] { "Cool", "Very nice", "Hate it" };
       int nextInt = new Random().nextInt(3);
       // save the new comment to the database
       comment = datasource.createComment(comments[nextInt]);
       adapter.add(comment);
-      break;
-    case R.id.delete:
-      if (getListAdapter().getCount() > 0) {
-        comment = (Comment) getListAdapter().getItem(0);
-        datasource.deleteComment(comment);
-        adapter.remove(comment);
-      }
-      break;
+    } else if (viewId == R.id.delete) {
+	      if (getListAdapter().getCount() > 0) {
+	        comment = (Comment) getListAdapter().getItem(0);
+	        datasource.deleteComment(comment);
+	        adapter.remove(comment);
+	      }
     }
     adapter.notifyDataSetChanged();
   }
